fix - pick next action based on policy & increase learning rate
replace pick_best_action with pick_action for picking the next action for
the next state when calculating the q table. this is because pick_action is
the actual policy we use for taking steps. in SARSA we need to use that to
calculate q values. using pick_best_action would make it off-policy.

also increased learning rate to 0.5 since 0.1 was too low, making it
hard to approach the true q values.